Cleaning AppDelegate of unused methods. UserDefaults wrapper. Rewrote Attraction Manager.
